Fixed multi-request batches; Client now submits batches (#9)
Fixes #6

Multi-request batches need subsequent requests to include the GUID that
was returned from prior server responses. To implement that,
EncodeTagBatchPart has been deprecated, and the client now submits
the request batches and interprets the server response.

Details:
This client doesn't properly handle multi-HTTP-request batches. The server responds with a GUID. When you have multiple parts, the second-Nth parts each need to include the batch guid to tell the server that they're all part of the same logical batch. That's what I added here.

I had to change the gohippo client signature to support this. Before, it would return an array of []byte - one record per serialized batch part. Well, we can't pre-serialize all the requests now, b/c we need to update the 2nd-Nth after receiving the 1st respond. So, I updated the client to actually submit the requests for the caller, and return an object that tells you how much it was able to accomplish.