Batch-building is now optimistic when oversized
When no populators are left that can fit under the desired size limit,
the BatchBuilder now tries one batch per populator, starting with the
smallest, rather than enforcing a hard limit on the request size. This
allows for possible success if the server can receive batches bigger
than the client's configured limit. If the populators increase in size
to the point that some will be rejected by the server, we've done the
best we can at submitting as much of the batch as possible.